+++ Lidgren.Network/NetSenderChannelBase.cs
@@ -12,8 +12,8 @@ namespace Lidgren.Network
 		internal abstract int GetAllowedSends();
 
 		internal abstract NetSendResult Enqueue(NetOutgoingMessage message);
-		internal abstract void SendQueuedMessages(float now);
+		internal abstract void SendQueuedMessages(double now);
 		internal abstract void Reset();
-		internal abstract void ReceiveAcknowledge(float now, int sequenceNumber);
+		internal abstract void ReceiveAcknowledge(double now, int sequenceNumber);
 	}
 }
